loading the incident data after clicking the load button in officer view

+++ app/Http/Controllers/IncidentController.php
@@ -4,37 +4,60 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Incident;
+use App\Models\Role;
+use App\Http\Controllers\HomeController;
+use Illuminate\Support\Facades\Auth;
 
 class IncidentController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
 
     protected function add()
     {
         $data=request()->all();
+        $user = Auth::user();
+        $role = Role::findOrFail($user->role_id);
+        $roleName = $role->roleName;      
+        if ($roleName == HomeController::CITIZEN){
+            Incident::create([
+                'incidentType_id' => $data['type'],                                    
+                'priority' => $data['options'],
+                'details' => $data['details'],
+                'lat' => $data['lat'],
+                'lng' => $data['lng']
+            ]);
 
-        //dd($data);
-        Incident::create([
-            'incidentType_id' => $data['type'],                                    
-            'priority' => $data['options'],//all the radio buttons named options
-            'details' => $data['details'],
-            'lat' => $data['lat'],
-            'lng' => $data['lng']
-        ]);
-
-        return view('thanks');
-
-        //after adding new incident you should view the list of incidents
+            return view('thanks');
+        }
     }
 
     protected function delete($incident){
 
-        $incidentObj=Incident::findOrFail($incident);
-        $incidentObj->delete();
+        $user = Auth::user();
+        $role = Role::findOrFail($user->role_id);
+        $roleName = $role->roleName;      
+        if ($roleName == HomeController::OFFICER){
+            
+            $incidentObj=Incident::findOrFail($incident);
+            $incidentObj->delete();
 
-        $incidents = Incident::all(); 
-        return view('officer',[
-            'incidents' => $incidents,
-        ]);    
+            $incidents = Incident::all(); 
+            return view('officer',[
+                'incidents' => $incidents,
+            ]);    
+
+        }
 
     }
+
+    protected function load($incident){
+        $incidentObj=Incident::findOrFail($incident);
+        return view ('incident-load',[
+            'incident' => $incidentObj
+        ]);
+    }
+
 }
